+++ Sources/OvEditor/src/OvEditor/Core/EditorResources.cpp
@@ -7,6 +7,7 @@
 #include <filesystem>
 
 #include <OvCore/Helpers/GUIDrawer.h>
+#include <OvCore/Helpers/GUIHelpers.h>
 #include <OvDebug/Assertion.h>
 #include <OvEditor/Core/EditorResources.h>
 #include <OvRendering/Settings/ETextureFilteringMode.h>
@@ -161,7 +162,7 @@ OvEditor::Core::EditorResources::EditorResources(const std::string& p_editorAsse
 	ValidateResources(m_shaders);
 
 	// Register the empty texture for the GUIDrawer to use it when a texture is missing
-	OvCore::Helpers::GUIDrawer::ProvideEmptyTexture(*m_textures["Empty_Texture"]);
+	OvCore::Helpers::GUIHelpers::ProvideEmptyTexture(*m_textures["Empty_Texture"]);
 }
 
 OvEditor::Core::EditorResources::~EditorResources()
